+++ photutils/segmentation/__init__.py
@@ -5,6 +5,7 @@ segmentation and measuring their centroids, photometry, and
 morphological properties.
 """
 
+from .catalog import *  # noqa
 from .core import *  # noqa
 from .deblend import *  # noqa
 from .detect import *  # noqa